lower pre upgrade node version

+++ upgrade.sh
@@ -38,20 +38,22 @@ npm set progress=false
 case $os in
     "Linux")
         if command -v dnf > /dev/null; then
-            echo "Updating Node"
+            if [[ "$node" < "12.13.1" ]]; then
+                echo "Updating Node"
 
-            curl -O https://nodejs.org/dist/v12.13.1/node-v12.13.1-linux-x64.tar.gz > /dev/null 2>&1
-            tar -xzf ./node-v12.13.1-linux-x64.tar.gz -C /usr --strip-components=1 --no-same-owner > /dev/null 2>&1
-            rm -f ./node-v12.13.1-linux-x64.tar.gz > /dev/null 2>&1
+                curl -O https://nodejs.org/dist/v12.13.1/node-v12.13.1-linux-x64.tar.gz > /dev/null 2>&1
+                tar -xzf ./node-v12.13.1-linux-x64.tar.gz -C /usr --strip-components=1 --no-same-owner > /dev/null 2>&1
+                rm -f ./node-v12.13.1-linux-x64.tar.gz > /dev/null 2>&1
 
-            node=$(node -v)
-            node=${node#"v"}
+                node=$(node -v)
+                node=${node#"v"}
 
-            echo "Node Updated to $node"
+                echo "Node Updated to $node"
 
-            source ~/.bashrc
+                source ~/.bashrc
 
-            sleep 0.2
+                sleep 0.2
+            fi
         elif command -v apt-get > /dev/null; then
             sleep 0.2
